requirements.txt
pandas==2.1.0
hydra-core==1.3.2
scikit-learn==1.3.0
matplotlib==3.8.0
statsmodels==0.14.0